smarter attack-positioning
Instead of just deciding whether to get closer to an enemy, the AI now also consideres all reachable tiles in a radius of 2 around itself and the target for whether it's better to attack from there.

This allows two things: Getting out of the close-quarter-combat-range to the target-unit and to pick the direction from which most damage can be dealt from.

The AI now is aware of the side an enemy is facing towards it and takes the respective armor into account for its decision-making about whom and from what angle to target an enemy.

For diagonal direction it will use the average of the two armors that it could be facing. For example if an enemy faces the unit with it's left front, the front-armor and left-armor are added up and the result divided by two.