Update AIModule.cpp
There is now a minimum-requirement for the quality of cover for hiding. If it can't be met while moving forwards, rather take a position with better cover than one closer to the enemy.

If no tile meets these minium criteria, choose the least bad tile available instead.

Fixed an issue that for the consideration of tiles to attack a target from, the check of whether an ally would be hit was done from the current tile instead of the tile that was considered to go to.