Revert to 6.3.3 logic
Reverted back to the high-risk-high-reward kind of playstyle for the AI, that it displayed in 6.3.3.
While it is disadvantageous in some situations, overall it just made for more dynamic and interesting play and put the pressure on the player. The recent changes simply made the AI too passive and thus less fun.