6.4.0
Added a new level of aggressiveness between balanced and reckless.
This level uses a less calculation-intense and less effective way to detect cover, which also makes it the recommended option for lower-end-computers.

For direct peaking there now is an additional check whether there would be a line of fire to the tile the enemy is supposed to be as opposed to just checking tile-visibility. This fixes an issue where the AI incorrectly assumed it would be able to spot a unit from a certain tile.