changed aggression 3
Aggression-option 3 is now no longer really much more aggressive than option 2. The way in which it differs is that it uses the Nodes on the map as well as the tiles that have adjecent cover and only does the cover-analysis for these instead of all reachable tiles.

This means it should still be quite capable of finding good cover and just not consider as many possibilities.