Running
AI-units that already see enemies will now act first. This should help against situations where the AI would sometimes overcommit by moving in a lot of units that aren't needed to kill an enemy.

When the option "Alternate movement methods" is enabled, the AI now also will make use of running given that it wouldn't run out of energy before running out of TUs. Since a lot of aliens have plenty of energy, prepare for them to make plenty of use of this ability.