6.6.1
Removed Aggressiveness-Option 3, that did something related to performance and unrelated to aggressiveness.
The functionality of what it did was moved to the option "Performance Optimisation".
"Performance Optimisation" is no longer enabled by default. You might want to enable it on weaker hardware or if you play mods that feature huge maps and lots of enemies.