performance improvements
The method that checked for the existence of proximity-grenades and whether reactionfire would be triggered no longer is called twice per tile but instead returns the other value as a reference-parameter. This results in a minor speed-gain.

Removed unneccessary line of sight checks before line of fire-checks. Not only did those make the turn-processing slower, they also lead to a bunch of false negatives and thus prevented otherwise possible attacks.

The distance to the closest tile to the target will now be determined strictly instead of also considering nearby tiles. In this case we already know the target is reachable, so there's no need to determine the closest alternative. This leads to a minor speed-gain.

With performance-optimisation enabled the calculations of how good a tile is to hide is no longer performed when the units looks for a tile to peek towards the enemy.

In any case, when the current tile is considered a peeking-tile the calculations to see whether it is a suitable hiding-tile will no longer be performed.

When looking for a hiding-tile the logic was reversed when it comes to checking first if a tile from where the unit could be seen from increased the threat or if the there is a line of sight. This prevents many unneccessary line-of-sight-checks and thus leads to a massive speed-gain.