New cover vs. attacking decisionmaking
The AI now considers taking cover befor deciding to attack. It will consider whether trying to hide is worth taking fewer shots or not even attempting to attack at all to remain hidden.
The quality of the potential cover plays a role on the balanced aggressiveness setting. On the lowest aggressiveness-setting taking cover will always be the most important thing.

Being near their friends no longer prevents the AI from prepriming their grenades if the other conditions for prepriming are met.

If an AI-unit has a grenade but the situation is not looking good for prepriming it, it will pro-actively look into creating a situation where it can savely prime a grenade. This includes not peeking and looking to go to the highest quality of cover.

Fixed an issue where for breaking line of sight the AI considered their own units standing in the way as breaking line of sight. While technically it was, it's not a particularly good idea to rely on your allies as cover.

AI now takes the No-LOS-accuracy-penalty into account.