7.0.2
For the decision of whether to apply the no-los-accuracy-penality the AI will now use the theoretical value instead of the actual visibility even when calculating that for the tile it currently stands on.

The reason for that is that when it used the actual visibility when facing away from the enemy it obviously would always fail and would think it had no line of sight on the target.