Update OptionsAdvancedState.cpp
There now is a new aggressiveness-value at the lower end of the spectrum, aggressiveness 0.
In this mode the AI will try to avoid being outdoors and it will not try to inch up on the player's units at all and instead exclusively look for how save they consider a hiding-spot to be. They will still peek out and attack when possible but usually most of them stay close to where they spawn until their enemies get near.