Update BattleUnit.cpp
The mapping of the aggression-stat to the Brutal-AI aggressiveness when using the inherit aggression-option was changed:
Aggression 0 and 1 map to their equivalents in brutal-AI-aggressiveness.
Everything else maps to 2.
This means the sweep-mode-behavior now only happens if you either manually set the aggressiveness to 3 or the unit has the LeeroyJenkins-flag set.