Update AIModule.cpp
Fixed a bracket-error in the code that decides whether a unit shall kneel down.

Fixed a bracket-error in the code that determines whether to reserve time-units for hiding.

Units controlled by Brutal-AI will now actively avoid light-sources in the darkness instead of ignoring their existence.

AI-units that would survive a proximity-grenade with more than half of their health, such as Lobstermen in Terror from the Deep, will now volunteer to walk over them in order to disarm them.

The AI now realizes that it is save to open a door if a proximity-grenade is one tile away from it and that they won't step onto it yet by doing so.

Fixed an issue that caused the code for checking how save a path is in terms of probable reaction-fire to only be run when the proximity-grenade-avoiding-logic was disabled.

Fixed an issue where blind-grenades were thrown at locations where targets were spotted several turns ago and likely long since have walked away instead of just when they were spotted during the last turn of the other faction.