Update Pathfinding.cpp
Units that use Brutal AI will no longer use Sneaky AI at the same time. The reason for that is that Sneaky AI can mess with the TU-costs and has a potential to cause more harm than good. Especially for melee-units that have to pass through other units line of sight to get closer.