Update RuleBaseFacility.cpp
The game no longer crashes with the "Not enough position vectors for crafts allocation."-message when a mod defines a hangar that can have several crafts but doesn't have sprite-offsets assigned to them.

Instead there will only be an error-message in the OpenXcom.log and all craft-sprites are drawn in the center of the hangar-sprite.

The proper way to use hangars that can hold several craft in Brutal-OXCE is to define an offset for the sprite drawing location from the hangar-sprite-center like this:
    crafts: 2
    craftSlots:
      - [-10, -10, 0]
      - [ 10, 10, 0]