new aggressiveness
There now is a new aggressiveness-level between the former levels 2 and 3.
The previous level 3 was moved to 4.
The new level 3 forces enemies to end their turn closer to the suspected position of the enemy than it started. It will try to find the position with the best cover that is eligible this way. When a unit thinks it was seen by the enemy this requirement must no longer be met and the unit is freed to hide where it thinks it's best.
Inherited aggressiveness from the unit's aggression now goes up to level 3.