tweaking attack- and hide after attack-logic
When performance-optimization is disabled, the AI will now run a complete targetability check when looking for tiles to attack from. This shall primarily help against all the false negatives that could force the AI to walk much further in order to attack or miss out on opportunities in the first place.

The AI will no longer gradually reposition to a better and better attack-position without doing a new check whether it can still hide. The gradual repositioning often lead to situations where a unit that intended to perform a hit&run-attack would not do it, as it longer had enough time-units.

AI units now take into account that when they come closer to perform an attack, they also need to preserve the time-units to go back to where they came from, if they decide to hide after attacking.