targetquality and origin-fixes
When looking for a location to attack from, the AI now takes into consideration that the origin-voxel depends on the facing of the unit, which once again depends on the location of the target relative to the potential attack-position.

This shall fix the false positives and false negatives when looking for positions to attack from. So if the AI looks for an attack-position it should now longer run into a situation where it then realizes it cannot actually attack from that position.

Smoke-peeking is now only done once per turn in order to prevent indiciveness when doing so.

Take into account that repositioning for an attack requires up to 4 time-units for turning towards the target after moving.

Targets in partial cover are now considered less attractive than fully exposed cover. This is realized by calculating a multiplier that is based on the ratio of voxels with a line of fire and maximum visible voxels of a unit.