Update AIModule.cpp
Fixed an issue with Autoplay where units would try to attack enemies that they don't see.

Fixed an inconsistenty with switching back and forth between wanting and not wanting to prime grenades.

Wanting to peak smoke now requires both the tile we go to and the suspected position of the enemy unit to be covered by smoke instead of just either of these.

If an enemy that was spotted nearby cannot be found within the time he could traverse along the edge of half the map, he will be assumed to be somewhere we haven't looked for a long time instead of the closest tile he could have reached.