behavioural changes
AI is no longer afraid of smoke.

There no longer is a separate peeking-algorithm for smoke. Instead the direct-peeking-algorithm takes the vision-range-reduction of smoke into account.

Fixed an issue where the AI went back and forth between wanting to go to a safer position in order to prime a grenade and then realizing it doesn't have enough time-units to do so and get closer again.

The AI no longer is afraid of light in the darkness. Note: This may have a slight negative impact on average playing-strength but having the AI actively avoid all light sources including soldiers could make finding them in base-missions too much of a nuisance. It also took away their own chance to attack the player because they positioned themselves too far away.

The hiding logic was revamped. Instead of trying to find a compromise between getting closer and being in a relatively save spot with convoluted additional conditionals, the AI will now always get closer as long as a spot provides maximum safety and always go to the savest spot available if no location is available that it considers perfectly safe.

The AI will no longer reserve TUs for hiding after shooting unless the hiding spot it could go to is considered perfectly safe.

Reverted a recent change that prevented the AI from reconsidering their attack-position again after something like finding new units or being reaction-shot at interrupted them on their path.

Positions near the attacking unit of the AI are no longer considered potential attack-positions. Only positions near the unit being attacked, where the attack-angle might be relevant to find a weak-spot in the armor.

The score for repositioning to attack a unit that explodes after death is now significantly reduced to counteract the willingness to get closer in order to have a higher chance of hitting.