Update AIModule.cpp
Fixed an issue that caused the check of whether hiding is more important than attacking to be skipped when a unit hadn't moved yet.

If a unit wants to hide with greater priority than attacking, it now takes into consideration that it may have to walk back to their previous spot.