restored pre 7.3.0 behavior but better attack vs. cover-logic
Fixed inconsistencies within the logic responsible for preserving time-units before considering to attack.

Cleaned up redundant code related to the time-unit-preservation-logic.

The amount of preserved time-units before going to a tile to attack is now determined under the premise that before the unit could attack, it didn't have a line of fire and thus the amount of time-units needed to make the last step are enough. This leads to much less attack-abortions.

Reverted 7.3.0 changes related to cover-evaluation. The AI now will once again try to find a compromise between cover and being close to the enemy instead of valuing cover so highly. Especially due to the rise in popularity of motion-scanners an overtly cautious approach has been proven ineffective.