Update AIModule.cpp
Fixed an issue that made the AI not consider positions at all for good and okay cover, when those also would qualify as great cover, when aggressiveness was 2 or higher.

Restored prime-position-seeking of 7.0.0 but changed it in a way that only positions are considered where there are enough time-units to perform the priming.

Reduced the impact the cuddle-avoid-modifier had on scoring of positions to roughly 1/5th of the previous impact. Before units often were pushed out to much less favorable positions due to wanting to avoid their proximity. It is only meant to spread out within positions of otherwise roughly equal evaluation.