peeking improvements
Indirect-peeking once again will prefer to check the first tile towards the supposed direction of the enemy that cannot be seen from the unit's position.

However, if that tile has already been explored this turn, the unit will determine alternative peek-positions based on the closest unexplored tile.

The AI will now realize when it can peek towards a location it wants to explore simply by turning.

The AI now can spend the same amount of time-units for peeking when it doesn't think it can reveal a unit as when it thinks it can, increasing the chance to discover their enemy.

Having a peek-score no longer invalidates a tile as potential cover.