Hack skipping marked soldiers in autocombat. They are simply skipped by the AI. Just a demonstration.
The turn automatically ends if all soldiers were either handled by the AI or had the Do not reselect button pressed.
There is some inconvenience at the end of a turn if it does not end automatically, which need to be fixed manually by disabling autocombat.
Also it might cause some endless looping of the AI between soldiers sometimes.

An extension for also controlling the AI of battle units instead of just soldiers, like mind controlled aliens, would not be hard.