final shaping
The logic to check whether a unit is AI-controlled or needs to be skipped in AI-handling is now done in a method that is part of the BattleUnit-object.

Fixed location of where autocombat is disabled when keep autocombat @ each battle is disabled to the constructor of BattleScapeGame rather than the init-method, which is at least once per turn.