hotfix
Units that use LOFTs other than a cylindrical one are now always treated as if they consisted of the biggest cylinder when it comes to hit- and visibility-calculations.
This should only impact units that were deliberately modded to be that way as the default is a cylinder with a radius of 3.