Fog of war-feature
New "Fog of war"-feature:
You can now enable "Fog of war" under Options => ADVANCED => Extended.
You can also select a color for the fog of war.

When enabled any tile that was previously visible to the player but currently isn't will now be drawn in the selected fog of war color.  The default is grey.