AI-improvements
AI unit's assumptions of whether they have been seen this turn no longer depends on whether they are next to a door. This had been a messy work-around based on doors giving away one's location, which no longer is the case.

Fatally wounded AI-units no longer become suicidal. It usually is still better to delay giving the enemy a kill.

When using AI-targetting-behaviour 1, the AI now is aware of it's limitations and no longer positions itself as if the limitation didn't exist just to then realize it can't attack and has wasted a bunch of time-units.

If an AI unit thinks it should not move and is already looking at the direction an enemy is expected to come from, it will now spin around in order to spot whether there's incoming enemies sneaking up from another direction.

Fixed a bug that caused the AI throw as many grenades as it could at a single target rather than none at all when "Allow AI to use explosives on turn one" was disabled and the turn from which they were allowed to use them wasn't reached yet.