Update AIModule.cpp
Using clearSight instead of hasTileSight to determine whether a tile can be seen from a certain position in the algorithm that determines a position from which the AI will take a peek.
Otherwise the AI would sometimes position it's units at spots from where they can only see a tiny part of a tile but couldn't see whether there's a unit on it.