aggressiveness-changes
Fixed an issue that could cause units of the AI to start searching a unit somewhere else when they actually knew where it was.

The aggressiveness-setting "4" no longer is the equivalent to an enabled LeeroyJenkins-flag.
It now works like this: Unseen units will still expend all their time-units to move closer to the enemy and not look for cover. However, when a unit thinks it has been seen by the enemy on its turn,it will consider taking cover. The routine that first checks whether there will be enough time-units to take cover after shooting is skipped. This means units with aggressiveness 4 will always attack as much as possible and only use leftover time-units to seek cover.

Units with a waypoint-weapon or psi-abilities will also skip the check timeunits for cover-check, since it is very well possible they can attack from where they are. This shall increase the frequency with which these kinds of attacks are being used by the AI.