new feature: intelligence
There is a new feature for Brutal-AI: Adjustable intelligence
This is split up between two options.
The first option allows you to pick a mode that determines what the intelligence of the enemies shall be based on:

Intelligence-mode
0> Static intelligence. 1> Inherit intelligence from unit-intelligence. 2> Inherit intelligence from difficulty-level.

The second allows you to pick an intelligence-level for the case static intelligence was selected as Intelligence-mode:

Static intelligence
Chance to make a smart move instead of a random one: 0> 0% 1> 20% 2> 40% 3> 60% 4> 80% 5> 100%

Internally there will be a random roll of 0-5. Whenever the value of the roll exceed the intelligence of the enemy, they will move to a random reachable tile incase their TUs are still full or skip moving alltogether if they already spent TU in this turn.
So at an intelligence of 0 the unit will always act randomly whereas at an intelligence of 5 it will never act randomly.

Note that other parts of the AI are not impaired by this. They will still turn towards where they think the enemy will come from and they will still relentlessly attack if they can. Movement that is part of the attack-routine currently is not impacted. This means melee-units that are in attack-range will be just as dangerous as before.