smart inventory
There is a new option in the OXCE-section:
"Smart ctrl-click and auto-equip"
It is enabled by default and has the following effect:
Ctrl-clicking items in the loadout-phase will prefer slots that are faster to reach. The off-hand will be deprioritized if the main-hand holds a two-handed item. Auto-equip uses the same logic.