no ground
Fixed an issue where certain items, like torpedoes in TFTD, would not be put into the backpack via ctrl-click-equip.
The issue was caused by considering the ground the item came from as possible slot to put it too and needing the same amount of TUs to take it into the hand from there.
The ground is now skipped as possible location to put items to.