aggressiveness 0 & 1 adjustment
When a unit controlled by the AI is set to use aggressiveness 0 or 1 and cannot find a hiding-spot that is considered as perfect, it will now go for the hiding spot that it thinks makes it most hidden as opposed to the best compromise between hidden and close to the enemy.

This shall make the behavior more distinct from aggressiveness 2.