further tweeks on aggressivnesses
On aggressiveness 0 the AI is no longer prevented from peeking.
When a unit thinks it has been seen in the current turn it is now allowed to put more preference in better hiding.
On aggressiveness 0 the hiding will now prefer to be as far away from the enemy as possible in addition to being well hidden.
On aggressivness 1 the AI will now also prefer hiding indoors.

Fixed an issue that sometimes prevented the AI from continuing to look around when it thought it should be able to see an enemy from it's current position but didn't due to smoke.