AI improvements
When realistic accuracy is enabled, the exposure is taken into account for determining positions to attack from.

AI no longer wastes time-units on peeking when in bug-hunt mode as there's no need to peek when you already know where the enemy is.

Fixed an issue where target-mode 4 was not taken into account for positioning.