ai changes
The AI now has a new heuristic to determine whether it should use one of their units as a spotter and which unit that should be.
This heuristic uses recent known enemy-positions to count how many time-units it could use offensively against these units if it had vision of them.
If the amount of time-units exceeds the amount of time-units a single unit has, the closest unit gets flagged as spotter and will walk up to the supposed enemy-positions until it sees a unit and thereby potentially sacrificing themselves.

The algorithm that makes the initial assumption of player-units being at starting-positions will now no longer think all units start on the same position and instead place them on tiles that are not used for other units yet.

Aggressiveness is no longer impacted by assumptions about the power-balance. At a certain point it was simply too suicidal with that.

Fixed an issue where units mind-controlled by the AI would think they are on the wrong team during movement.

Aggressiveness no longer impacts the cuddle-avoid-modifier. Once again for the reason of becoming too suicidal.

Indoor-tiles now provide a relatively small flat score-bonus instead of a scaling one based on aggression.

Having a line of sight to corpses or doors now reduces the attractiveness of tiles for the purpose of hiding. The more doors and corpses the less attractive the tile is considered.

AI-units will no longer want to end their turn early if their intelligence is 5 or higher.

Autoplay is now fixed to play at maximized aggressiveness.