intelligence rework
Reworked how intelligence impacts the behaviour of units controlled by Brutal-AI since the previous iteration had too little impact whereas the one before that had too much.

So the new iteration is something in-between.

A first roll will determine whether to fully randomize the score for the current action or to do second roll where a weighted random multiplier is used.
The chance to fully randomize scores is (5 - intelligence) * 20%.

The multiplier is only applied to moves that have a score of more than 0. This way completely ridiculous moves are filtered out and the resulting moves will still somewhat fulfill their purpose.