new option
There is a new advanced battle-scape-option in BOXCE:

Precise projectile-tracking => Allows both the player, when holding the alt-key, and Brutal-AI to tell the exact tile an enemy projectile was fired from without having seen the unit that fired.

This was previously always the case but now can be turned off. Disabling it will help units with stealth- and vision-advantage to rely on reaction-fire-based-strategies without instantly triggering retaliation to their exact position.
The AI will still know where the shot came from but can no longer target that tile without visual confirmation.
The player will no longer be able to see the location a shot came from by holding the alt-key.