fixes
Fixed the AI ignoring the penalty for not having line of sight, when it was set to always miss.
The AI no longer will consider the penalty for not having line of sight when it comes to positioning itself for attacking.
The AI no longer thinks it is outside of the vision range of enemy units when the enemy units are outside of the AIs vision-range.