AI-improvements
Brutal AI is now capable of replacing their current weapon with a better one from a dead nearby friend or enemy if the sitation looks calm enough to do so unharmed.

When a unit went out of their way to visit a corpse for a better weapon and it looks save enough, they will also pick up additional ammunition and grenades as long as they have enough time-units and it would not lead to encumberance.

AI-controlled units with an aggression of 1 or higher are now forced to get closer to the assumed position of an enemy unless they think they were spotted in the current turn. This means they will be much more proactive and should always try to hunt the player down.

The AI will now consider more opportunities as a good enough to pre-prime grenades.