Update BattlescapeGame.cpp
Reverted a change that caused the AI to consider units who still have grenades as armed in the sense of not looking to find an additional weapon as it did more harm than good.

Fixed an issue that caused the AI to look for a new weapon and throw away their current one, when it wasn't loaded despite still having ammo for that weapon in it's inventory when it didn't have enough TUs to reload the weapon in the same turn. This would happen particularly often with 1-shot-weapons.