Update AIModule.cpp
If a weapon causes negative damage to armor, morale, energy, mana or time-units, this effect is now ignored in the damage-type-modifier of the AI-score for using that weapon.

Incase you're wondering why this is worthy to be done: In X-Piratez there's a weapon called "Handle". In addition to it's main-effect of stunning enemies, it for some reason also restored a ton of morale for them. It restored so much morale, that the AI thought this outweighs the stun-damage and didn't want to use it.

So pro-tip for X-Piratez: Hit your own units with a handle to restore their morale!