Update AIModule.cpp
Removed spotter-logic as this could be exploited to get enemies stranded in bad positions without accomplishing anything.

Fixed various causes of units not looking in the right direction when they ended their turns.

Fixed an issue with the peeking-logic that could cause going back and forth to the same tile to look into different directions from there instead of just turning when already standing on the tile.

Reverted prior change of making AI more cover-oriented in aggressiveness 1.

Reverted prior change of forcing AI to alway move forward in aggressiveness 1.