Update AIModule.cpp
AI peeking-logic has been revamped.
AI units will now go to a tile within peeking-range from where there is a line of sight to the maximum amount of tiles belong to a path that leads to the position of where the closest assumed enemy unit is.
Once this tile is reached it will look into all directions in which it will uncover tiles it hasn't seen this turn.
This will, for example, allow the AI to look out of windows much more than it did before. Especially in combination with the new "Camper"-aggressiveness-level.

The AI no longer predicts reaction-fire via visibility-checks from assumed enemy-unit's positions. The algorithm that did this was very processing-intense and could cause extremely long turn-times if flying units with a lot of time-units were checking the safety for a lot of possible paths.
Instead the AI will now simply use it's assumption of whether it is seen by the enemy to let it's decision-making about where to move be impacted.
While this change will reduce playing-strength, it was deemed necessary due to the extreme waiting-times. If you have been using AI-performance optimization before it may no longer be necessary.

AI performance-optimization won't bother to analyze the validity of tiles that are at least 2 levels above the closest floor and have no ceiling for any purpose including attacking.

AI will ignore the cuddle-avoid-modifier, grenades and fire for peeking since it's very likely it'll move after peeking anyways and would just be wasting time-units for taking these things into consideration.

When a discovered unit tries to hide it will now prefer higher armor when hiding indoors.

AI-units with lower intelligence than 5 will no longer just end their turn despite seeing an enemy unit. That wasn't intended and while they should be less smart, they weren't supposed to be that dumb.