turn-logic-revamp
The decision-making of whether a unit will turn to look around is now handled before the movement-logic code.
The AI is slightly smarter about when and when not to expend time-units for looking around.
This saves a lot of unnecessary processing-time and allows stationary units like turrets to look around too.
Fixed an issue that could cause an AI unit not to look where it's supposed to look.
Removed code-relics from when the AI had no concept of making assumptions on enemy-whereabouts. The only effect that had was that they kept looking for enemies when none were alive anymore.
Only AI units who have not yet revealed themselves to their enemy will consider not attacking when they can in order to go to a better hiding-spot.