grenading
When looking for a position to attack from, a unit with a pre-primed grenade will now also be okay with a position with no line of fire as long as it's possible to throw the grenade.

Fixed an issue where a unit wouldn't correctly consider whether they are standing next to a door when deciding whether to continue peeking. This lead to peeking when they should not and not peeking when they should.

Even at intelligence 5 or higher the AI will now want to end the turn of a unit that went to a cover-spot.