Fixed discover-threat-issue
When determining the likelyness of being discovered on a certain tile the AI looked at the potential paths their enemies could go from the locations they were assumed to be. However, the pathfinding that did this could sometimes fail due to the units actual position being in the way of the unit's assumed position and the unit not wanting to walk through itself.
This is fixed now. What this actually fixes is that AI units sometimes were very reckless and exposing themselves to be found by the player when they shouldn't. This was especally prevalent in narrow areas.