Update AIModule.cpp
When a unit would call of an attack in order to hide it now checks whether there actually is any location it might want to hide at. This prevents an issue where a unit just walks forward without attacking or hiding when it could.

When bug-hunt-mode for AI is enabled it will no longer waste TUs on searching enemies.