Update AIModule.cpp
The AI's peaking-algorithm will now try to get the best compromise between newly discovered tiles and time-units-preserverd instead of looking for the longest part of the path towards the assumed position of the closest enemy.

This shall help spotting units that try to sneak up on them from behind while also preserving more TUs for reaction-shots.