AI adjustments
For missions that have an objective, missions that have another level and base-defense-missions the AI will now play with drastically increased aggressiveness.

Improved the logic for when the AI will turn around on the spot to only do it, when the unit knows there is someone to find.

Fixed a bug in the check about whether the unit would drop down somewhere while scouting.

The AI will no longer go to a spot from which it can throw a grenade when the target it wants to throw at already stands in the blast-radius of another grenade.

AI-units will now go to a tile from which most tiles can be seen while looking in only one direction.
They also will then look into that direction.

When indoors the AI is more willing to spend more time-units for scouting. When outdoors the AI will try to preserve more time-units for getting into better cover.

AI units that have showed themselves to the enemy will no longer get closer if they have no TUs to hide or attack. Instead they will try to increase distance. This does not apply to units who's weapon-range would be too short to walk away.