AI-fixes
Fixed an issue where AI-units wouldn't check again if they can attack after turning on the spot to scout their surroundings. This lead to them moving unnecessarily and getting reaction-shot.

For the calculation of smoke's vision-impairing effects, the AI now uses the worst-case scenario instead of an often too optimistic average. This shall help them to scout more aggressively for units hiding inside smoke.

Increased the negative effect of corpses in the line of sight on the score for hiding-spots.

Elevator-tiles now also get their score reduced for being hiding-spots.