Update AIModule.cpp
If there is a great-cover spot that is closer to the assumed enemy-position than the current position, indirect peeking is skipped on aggressivenesses greater than 0.
This should lead to a bigger behavioural distinction between aggressiveness 0 and 1 in the sense that 1 brings units to the "frontline" more quickly while 0 remains more territorial.