Update AIModule.cpp
On Aggressiveness 0 and 1 the cover-detection-algorithm now also is used for tiles that are within the enemie's line of fire.
Afterall it is definitely better to expose yourself to only one enemy than going for a potentially much worse "fallback"-tile that could otherwise be the result of not checking for best possible cover.
This will primarly help AI-units that cannot break line of sight to everyone after they have attacked and don't have enough TUs to go for a real cover.